Fixed style data type
git-svn-id: https://logicmail.svn.sourceforge.net/svnroot/logicmail/trunk@756 5c734088-3d25-0410-9155-b3c3832efda5

+++ LogicMail/src/org/logicprobe/LogicMail/ui/ActiveFieldManager.java
@@ -66,7 +66,7 @@ public class ActiveFieldManager extends VerticalFieldManager {
      *
      * @param style Styles(s) for this manager.
      */
-    public ActiveFieldManager(int style) {
+    public ActiveFieldManager(long style) {
         super(style);
     }
